requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==1.7.0
Django==4.1.2
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.0rc9
iniconfig==1.1.1
mysqlclient==2.1.1
packaging==21.3
pluggy==1.0.0
pycodestyle==2.9.1
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.5
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
